fix(HostReports): Separate general/historical views
